Remove createServer function
This refactor added little value.
