--- server/createServer.js
@@ -1,44 +1,0 @@
-import express from 'express'
-import webpack from 'webpack'
-import requestLanguage from 'express-request-language'
-import webpackDevMiddleware from 'webpack-dev-middleware'
-import webpackHotMiddleware from 'webpack-hot-middleware'
-import webpackHotServerMiddleware from 'webpack-hot-server-middleware'
-import clientConfig from '../webpack/client.dev'
-import serverConfig from '../webpack/server.dev'
-
-export const createServer = () => {
-  const DEV = process.env.NODE_ENV === 'development'
-  const publicPath = clientConfig.output.publicPath
-  const outputPath = clientConfig.output.path
-  const app = express()
-
-  app.use(
-    requestLanguage({
-      languages: ['en', 'fr'],
-    }),
-  )
-
-  // UNIVERSAL HMR + STATS HANDLING GOODNESS:
-  if (DEV) {
-    const multiCompiler = webpack([clientConfig, serverConfig])
-    const clientCompiler = multiCompiler.compilers[0]
-
-    app.use(webpackDevMiddleware(multiCompiler, { publicPath }))
-    app.use(webpackHotMiddleware(clientCompiler))
-    app.use(
-      // keeps serverRender updated with arg: { clientStats, outputPath }
-      webpackHotServerMiddleware(multiCompiler, {
-        serverRendererOptions: { outputPath },
-      }),
-    )
-  } else {
-    const clientStats = require('../buildClient/stats.json') // eslint-disable-line import/no-unresolved
-    const serverRender = require('../buildServer/main.js').default // eslint-disable-line import/no-unresolved
-
-    app.use(publicPath, express.static(outputPath))
-    app.use(serverRender({ clientStats, outputPath }))
-  }
-
-  return app
-}
